fix(deps): update tree-sitter packages

+++ src/analyzers/repomap.rs
@@ -492,7 +492,7 @@ fn get_call_node_kinds(lang: Language) -> Vec<&'static str> {
 /// Extract the function name from a call expression.
 fn extract_call_name(node: &tree_sitter::Node<'_>, source: &[u8]) -> Option<String> {
     // Try to find the function/identifier node
-    for i in 0..node.child_count() {
+    for i in 0..node.child_count() as u32 {
         if let Some(child) = node.child(i) {
             let kind = child.kind();
             if kind == "identifier"
@@ -509,7 +509,7 @@ fn extract_call_name(node: &tree_sitter::Node<'_>, source: &[u8]) -> Option<Stri
                 if let Some(right) = child
                     .child_by_field_name("field")
                     .or_else(|| child.child_by_field_name("property"))
-                    .or_else(|| child.child(child.child_count().saturating_sub(1)))
+                    .or_else(|| child.child(child.child_count().saturating_sub(1) as u32))
                 {
                     let text = right.utf8_text(source).ok()?;
                     return Some(text.to_string());
